Remove redundant build from .travis.yml file.
The `build` in the `.travis.yml` file is redundant. `stack test` is just
an alias for `stack build --test`.

Also in the old way of doing things,  `-Werror` is actually not set when
running tests. So it is ignored for  test files..

Now also test files need to have no warnings.