Remove `ToSample a => ToSample [a]` instance
`ToSample [a]` was generating an infinite list of samples. This is far from
being useful; if you have two samples for `a` servant-docs generates 79kloc long
samples for `[a]`.

Because it's impossible to come up with a useful `ToSample [a]` that works for
all use cases, we leave that out.