Don't build --dep for all
As tutorial depends on servant-js, which isn't in this repository
cabal fails fatally to `new-build --dep all`.

Instead we are building deps for servant-server and servant-client only,
which already includes e.g. http-client and warp, i.e. pretty much
everything we need. It's not bad that some dependencies might be build
during `script` phase