Remove redundant point
# Remove redundant point
There are so much waste in the universe and we need to change the situation. The better usage the better habits of one leads a positive attitude of everyone.
# Improve disk space usage
You always hear guys complain about the cost of storage so it's appreciated as fuck. Decreasing line by 1 symbol leads to less disk space usage (up to 2 bytes at a time)!