uverb: Remove To/FromJSON instance completely
I just realised it's not needed at all! Servant only needs the
MimeRender typeclass to work.

This gets rid of the overlapping instance between
  MimeRender ctype (WithStatus a) and
  ToJSON a => MimeRender JSON a

WithStatus does not need a (To/FromJSON) instance, as the MimeRender
instance we had makes use of the _underlying_ To/FromJSON instance
already.

Even nicer is that WithStatus now needs no additional instances to
function.  It will work for any content type!