WIP: Add prometheus metrics to the servant router
Annotates the routing tree with a middleware that instruments each endpoint in your servant application.
For each endpoint we will keep track of a histogram with request latency + request count

Captures are automatically filtered; such that  /users/13813 and /users/24774
get mapped to the same metric.

Unfortunately the name of the capture is not preserved; as we throw away that
data when we're at the point of routing. We could perhaps change this but it's
more significant change.