feat: Polymorphic `Elem` for `Union`
The `Elem` call was not possible to resolve with polymophic type. For
example, the following `forall a. Elem (Maybe a) '[Maybe a, Int]` cannot
be resolved.

I don't really understand the reasons why, looks like the former
implementation with `If (a == b)` does not work with polymophic types.

The new implementation uses pattern matching in closed type families.

Close #1590 and #1381.